pyproject.toml
[tool.poetry]
name = "search-engine"
version = "0.1.0"
description = ""
authors = ["TimurMeshchenko <timur_meshchenko@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
beautifulsoup4 = "^4.12.3"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
